% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchLinkPlot.R
\name{BatchLinkPlot}
\alias{BatchLinkPlot}
\title{Title}
\usage{
BatchLinkPlot(
  seurat.obj1,
  seurat.obj2,
  cell4plot.order = NULL,
  cell4plot.color = c("#BC3C29", "#A1433D", "#864B51", "#6B5365", "#505A79", "#35628D",
    "#1A6AA1", "#0072B5", "#2075A0", "#40788C", "#607B78", "#807E63", "#A0814F",
    "#C0833B", "#E18727", "#C5862C", "#A98632", "#8E8637", "#72853D", "#578542",
    "#3B8548", "#20854E", "#2C825C", "#39806A", "#457E78", "#527C86", "#5E7A94",
    "#6B78A2", "#7876B1", "#767BB0", "#7580AF", "#7485AF", "#728AAE", "#718FAE",
    "#7094AD", "#6F99AD", "#83A2A9", "#98ACA5", "#ACB5A1", "#C1BF9D", "#D5C899",
    "#EAD295", "#FFDC91", "#FCC791", "#FAB292", "#F79E93", 
     "#F58994", "#F27595",
    "#F06096", "#EE4C97"),
  save.format = "pdf"
)
}
\arguments{
\item{seurat.obj1}{the first Seurat object with completed dimensionality reduction analysis}

\item{seurat.obj2}{the second Seurat object with completed dimensionality reduction analysis}

\item{cell4plot.order}{specify the order of clusters for cell4plot}

\item{cell4plot.color}{specify the color of clusters for cell4plot}

\item{save.format}{specify the format for saving the plot}
}
\value{
return a list of plots for all clusters
}
\description{
Title
}
\examples{
brain.harmony$cell4plot = brain.harmony$orig.ident
BatchLinkPlot(
    seurat.obj1 = brain.harmony,
    seurat.obj2 = brain.cca,
    cell4plot.order = c("cortex", "hippocampus"),
    save.format = "png"
    )

}
