% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchProcrustesPlot.R
\name{BatchProcrustesPlot}
\alias{BatchProcrustesPlot}
\title{Title}
\arguments{
\item{seurat.obj1}{the first Seurat object with completed dimensionality reduction analysis}

\item{seurat.obj2}{the second Seurat object with completed dimensionality reduction analysis}

\item{DR.method}{specify a common dimensional reduction method for two Seurat objects (e.g., umap, tsne, pca)}
}
\description{
Title
}
\examples{
brain.harmony$cell4link = brain.harmony$orig.ident
BatchProcrustesPlot(seurat.obj1 = brain.harmony, seurat.obj2 = brain.cca, DR.method = "pca")

}
